requirements.txt
Django
psycopg2
freeze
djangorestframework
django-filter 
djangorestframework-jwt
